Extract the script from MakersNewView.vue into a composable
