+++ frontend/src/components/makers/MakersNewView.vue
@@ -1,74 +1,32 @@
 <script setup lang="ts">
-import { ref, onMounted } from 'vue'
-import axios from 'axios'
-import { useRouter } from 'vue-router'
-import { checkLoginStatus } from '@/components/utils.js'
-
-interface MakerResponse {
-  id: number
-  name: string
-  postal_code: string
-  address: string
-  phone_number: string
-  fax_number: string
-  email: string
-  home_page: string
-  manufacturer_rep: string
-}
+import { onMounted } from 'vue'
+import { useMakers } from '@/composables/useMakers.ts'
 
 interface MessageEvent {
   type: 'danger' | 'success' | 'warning' | 'info'
   text: string
 }
 
-const API_BASE_URL = import.meta.env.VITE_API_BASE_URL as string
-
 const emit = defineEmits<{
   message: [payload: MessageEvent]
 }>()
 
-const router = useRouter()
-
-const maker = ref<MakerResponse | null>(null)
-const name = ref<string>('')
-const postalCode = ref<string>('')
-const address = ref<string>('')
-const phoneNumber = ref<string>('')
-const faxNumber = ref<string>('')
-const email = ref<string>('')
-const homePage = ref<string>('')
-const manufacturerRep = ref<string>('')
-const errorMessage = ref<string>('')
-
-const makerRegistration = async (): Promise<void> => {
-  try {
-    const response = await axios.post<MakerResponse>(`${API_BASE_URL}/makers`, {
-      maker: {
-        name: name.value,
-        postal_code: postalCode.value,
-        address: address.value,
-        phone_number: phoneNumber.value,
-        fax_number: faxNumber.value,
-        email: email.value,
-        home_page: homePage.value,
-        manufacturer_rep: manufacturerRep.value
-      }
-    })
-    maker.value = response.data
-    emit('message', { type: 'success', text: 'メーカー情報を1件登録しました。' })
-    router.push(`/makers/${maker.value.id}`)
-  } catch(error) {
-    if (axios.isAxiosError(error) && error.response?.status === 422) {
-      errorMessage.value = '入力に不備があります。'
-    }
-  }
-}
+const {
+  name,
+  postalCode,
+  address,
+  phoneNumber,
+  faxNumber,
+  email,
+  homePage,
+  manufacturerRep,
+  errorMessage,
+  makerRegistration,
+  loggedIn
+} = useMakers(emit)
 
 onMounted(async (): Promise<void> => {
-  await checkLoginStatus(() => {
-    emit('message', { type: 'danger', text: 'ログインが必要です。' })
-    router.push('/')
-  })
+  await loggedIn
 })
 </script>
